Reintroduce float math
