+++ examples/ext-math.cf
@@ -1,5 +1,4 @@
-~require-ext-math
-
+~1 echo c!
 $2.23 $3.45 ~f+ fdup
 f.s f.
 fdup f.s f* f.
@@ -16,14 +15,14 @@ $1.1 $2.2 $3.3 ~f.s frot f.s f-rot f.s
 fnip f.s
 
 
-~float alloc dup $11.11 ~swap f! f.s
-f@ f.s
+~\ float alloc dup $11.11 ~swap f! f.s
+~f@ f.s
 
-:test ~$13.3 ~[f,] ^f@ f. ,;
+:test ~$13.3 ~[f,] ^f@ f. ;
 
 ~test test
 
-:test-add ~$12.3 ~[f,] $2.5 ~[f,] ^f@ f@ f+ f. ,;
+:test-add ~$12.3 ~[f,] $2.5 ~[f,] ^f@ f@ f+ f. ;
 
 ~test-add test-add
