--- ext/ext-math.h
@@ -1,18 +1,0 @@
-// The author disclaims copyright to this source code.
-#ifndef __EXT_MATH_H
-#define  __EXT_MATH_H
-
-#ifdef __EXTENDED_MATH
-
-struct fstack
-{
-  // stack
-  number_t *cells;
-  // stack position
-  int sp;
-  int lim;
-};
-
-#endif
-
-#endif /* __EXT_MATH_H */
