+++ src/core.c
@@ -1,15 +1,31 @@
 // The author disclaims copyright to this source code.
 
 /**********************************************************************************
+ *   PREFIX DEF
+ *********************************************************************************/
+#ifdef __SECTION_PREFIX_DEF
+
+case ':': { s->color = define; echo_color(s, c, COLOR_RED); return; }
+case '^': { s->color = compile; echo_color(s, c, COLOR_GREEN); return; }
+case '~': { s->color = execute; echo_color(s, c, COLOR_YELLOW); return; }
+case '\'': {
+  s->color = (s->color == execute || s->color == tick || s->color == tick_entry) ? tick : compile_tick;
+  echo_color(s, c, COLOR_BLUE);
+  return;
+}
+case '`': {
+  s->color = (s->color == execute || s->color == tick || s->color == tick_entry) ? tick_entry : compile_tick_entry;
+  echo_color(s, c, COLOR_BLUE);
+  return;
+}
+
+#endif /* __SECTION_PREFIX_DEF */
+
+
+/**********************************************************************************
  *   HASH DEF
  *********************************************************************************/
 #ifdef __SECTION_HASH_DEF
-
-#ifdef __KEEP_ENTRY_NAMES
-#define ENTRY_NAME(x) (x)
-#else
-#define ENTRY_NAME(x) (NULL)
-#endif
 
 // Define __SHOW_MISSING_HASH in conf.h to display name <-> hash relationship
 //
